_posts/: refresh articles and condensate information
This article is being reached massively from the internet,
but the info is spreaded into two articles. The main landing post
from search engines is the one which contains less information.

Let's ease the think for users, refreshing and condensating the information.
We end now with a single article about this issue.

Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>